% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stack}
\alias{stack}
\title{Create a stack}
\usage{
stack(init = 20)
}
\arguments{
\item{init}{Initial size of the list that backs the stack. This is also used
as the minimum size of the list; it will not shrink any smaller.}
}
\description{
A stack object is backed by a list. The backing list will grow or shrink as
the stack changes in size.
}
