% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastmap.R
\name{fastmap}
\alias{fastmap}
\title{Create a fastmap object}
\usage{
fastmap(missing = NULL)
}
\arguments{
\item{missing}{The value to return when \code{get()} is called with a key
that is not in the map. The default is \code{NULL}, but in some cases it
can be useful to return a sentinel value, such as a
\code{\link{key_missing}} object.}
}
\description{
Create a fastmap object
}
