% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_missing.R
\name{key_missing}
\alias{key_missing}
\alias{is.key_missing}
\title{A missing key object}
\usage{
key_missing()

is.key_missing(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
A \code{key_missing} object represents a missing key.
}
